requirements.txt
attrs
psycopg2
sqlalchemy